+++ mt_metadata/common/geographic_location.py
@@ -10,11 +10,9 @@ from pydantic import Field
 # =====================================================
 class GeographicLocation(MetadataBase):
     country: Annotated[
-        str | None,
+        str | list[str] | None,
         Field(
             default=None,
-            type="string",
-            items={"type": "string"},
             description="Country of the geographic location, should be spelled out in full. Can be a list of countries.",
             examples="United States of America",
             alias=None,
@@ -23,14 +21,12 @@ class GeographicLocation(MetadataBase):
                 "required": False,
             },
         ),
-    ] = None
+    ]
 
     state: Annotated[
-        str | None,
+        str | list[str] | None,
         Field(
             default=None,
-            type="string",
-            items={"type": "string"},
             description="State or province of the geographic location, should be spelled out in full. Can be a list of states or provinces.",
             examples="[Colorado, Utah]",
             alias=None,
@@ -39,14 +35,12 @@ class GeographicLocation(MetadataBase):
                 "required": False,
             },
         ),
-    ] = None
+    ]
 
     county: Annotated[
-        str | None,
+        str | list[str] | None,
         Field(
             default=None,
-            type="string",
-            items={"type": "string"},
             description="County of the geographic location, should be spelled out in full. Can be a list of counties.",
             examples="[Douglass, Fayet]",
             alias=None,
@@ -55,64 +49,60 @@ class GeographicLocation(MetadataBase):
                 "required": False,
             },
         ),
-    ] = None
+    ]
 
     township: Annotated[
-        str | None,
+        str | list[str] | None,
         Field(
             default=None,
             description="Township or city name or code.",
             examples="090",
-            type="string",
             alias=None,
             json_schema_extra={
                 "units": None,
                 "required": False,
             },
         ),
-    ] = None
+    ]
 
     section: Annotated[
-        str | None,
+        str | list[str] | None,
         Field(
             default=None,
             description="Section name or code.",
             examples="012",
-            type="string",
             alias=None,
             json_schema_extra={
                 "units": None,
                 "required": False,
             },
         ),
-    ] = None
+    ]
 
     quarter: Annotated[
-        str | None,
+        str | list[str] | None,
         Field(
             default=None,
             description="Quarter section code.",
             examples="400",
-            type="string",
             alias=None,
             json_schema_extra={
                 "units": None,
                 "required": False,
             },
         ),
-    ] = None
+    ]
 
     parcel: Annotated[
-        str | None,
+        str | list[str] | None,
         Field(
             default=None,
             description="Land parcel ID.",
             examples="46b29a",
-            type="string",
             alias=None,
             json_schema_extra={
                 "units": None,
                 "required": False,
             },
         ),
-    ] = None
+    ]