+++ mt_metadata/timeseries/station_basemodel.py
@@ -1,21 +1,26 @@
 # =====================================================
 # Imports
 # =====================================================
-
+from enum import Enum
 from typing import Annotated
 
+from pydantic import Field, ValidationInfo, field_validator
+
 from mt_metadata.base import MetadataBase
-from pydantic import Field
 from mt_metadata.common import (
+    Comment,
     DataTypeEnum,
     ChannelLayoutEnum,
+    StationLocation,
+    Copyright,
     Fdsn,
-    Location,
+    Instrument,
+    Provenance,
+    TimePeriod,
     Person,
-    Declination,
-    GeographicLocation,
+    Orientation,
 )
-
+from mt_metadata.utils.list_dict import ListDict
 
 # =====================================================
 
@@ -27,20 +32,18 @@ class Station(MetadataBase):
             default=None,
             description="How the station channels were laid out.",
             examples="x",
-            type="string",
             alias=None,
             json_schema_extra={
                 "units": None,
                 "required": False,
             },
         ),
-    ] = None
+    ]
 
     channels_recorded: Annotated[
         str,
         Field(
             default=[],
-            type="string",
             items={"type": "string"},
             description="List of components recorded by the station. Should be a summary of all channels recorded. Dropped channels will be recorded in Run metadata.",
             examples='"[ Ex, Ey, Hx, Hy, Hz, T]"',
@@ -52,28 +55,13 @@ class Station(MetadataBase):
         ),
     ]
 
-    comments: Annotated[
-        str | None,
-        Field(
-            default=None,
-            description="Any comments on the station.",
-            examples="5 runs",
-            type="string",
-            alias=None,
-            json_schema_extra={
-                "units": None,
-                "required": False,
-            },
-        ),
-    ] = None
-
+    comments: Comment
     data_type: Annotated[
         DataTypeEnum,
         Field(
             default=BBMT,
             description="Type of data recorded. If multiple types input as a comma separated list.",
             examples="BBMT",
-            type="string",
             alias=None,
             json_schema_extra={
                 "units": None,
@@ -88,7 +76,6 @@ class Station(MetadataBase):
             default="",
             description="Closest geographic name to the station, usually a city, but could be another common geographic location.",
             examples='"Whitehorse, YK"',
-            type="string",
             alias=None,
             json_schema_extra={
                 "units": None,
@@ -103,7 +90,6 @@ class Station(MetadataBase):
             default="",
             description="Station ID name.  This should be an alpha numeric name that is typically 5-6 characters long.  Commonly the project name in 2 or 3 letters and the station number.",
             examples="MT001",
-            type="string",
             alias=None,
             pattern="^[a-zA-Z0-9]*$",
             json_schema_extra={
@@ -117,7 +103,6 @@ class Station(MetadataBase):
         str,
         Field(
             default=[],
-            type="string",
             items={"type": "string"},
             description="List of runs recorded by the station. Should be a summary of all runs recorded.",
             examples='"[ mt001a, mt001b, mt001c ]"',
@@ -128,3 +113,10 @@ class Station(MetadataBase):
             },
         ),
     ]
+
+    @field_validator("comments", mode="before")
+    @classmethod
+    def validate_comments(cls, value, info: ValidationInfo) -> Comment:
+        if isinstance(value, str):
+            return Comment(value=value)
+        return value